refactor: move deck import types to model
