+++ src/features/import/hooks/useDeckImport.ts
@@ -19,7 +19,7 @@ import {
 } from "@/entities/card";
 import { createDeck, generateDeckId, useDeckMutations, useDecks } from "@/entities/deck";
 import { useSession } from "@/entities/session";
-import type { DeckImportPreview, DeckImportResult, DeckImportRow } from "@/features/import/components/deckImportTypes";
+import type { DeckImportPreview, DeckImportResult, DeckImportRow } from "@/features/import/model/deckImportTypes";
 import { parseCsv } from "@/features/import/lib/cardCsv";
 import { buildDeckImportPlan } from "@/features/import/lib/deckImportAnalysis";
 import sampleCards from "../../../../sample/build/output.json";
