+++ src/features/import/index.ts
@@ -1,4 +1,4 @@
-export type { DeckImportPreview, DeckImportResult } from "./components/deckImportTypes";
+export type { DeckImportPreview, DeckImportResult } from "./model/deckImportTypes";
 export { useDeckImport } from "./hooks/useDeckImport";
 export { useSampleDeckBootstrap } from "./hooks/useSampleDeckBootstrap";
 export { downloadSampleCsv, SAMPLE_CSV_TEXT } from "./sampleCsv";
